[DEBT-043-D7] Split agent/providers/llm.py into focused modules
- Split llm.py (471 LoC) into llm/ package with 6 modules:
  - __init__.py (22): Public API re-exports
  - _pricing.py (69): Token cost estimation and pricing
  - _prompts.py (43): System and analysis prompt templates
  - _schemas.py (74): Pydantic models and Protocol definitions
  - _claude.py (244): ClaudeSynthesizer implementation
  - _mock.py (68): MockSynthesizer for testing
  - _factory.py (40): get_synthesizer factory function
- All files now ≤400 lines (largest: _claude.py at 244)
- Added TC001 exemption for _schemas.py (Pydantic model needs runtime import)
- Quality gates: pre-commit, pytest (1052 passed)
